requirements.txt
cffi==2.0.0
cryptography==46.0.5
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
